backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
python-dotenv
sqlalchemy[asyncio]
sqlmodel
asyncpg
psycopg2-binary
psycopg
psycopg-binary
python-jose[cryptography]
passlib[bcrypt]
supabase
httpx
pytest
pytest-asyncio
aiofiles
python-multipart
email-validator
requests
beautifulsoup4
lxml
PyYAML
setuptools
invoice2data
PyMuPDF
gunicorn
cryptography
alibabacloud-ocr-api20210707
alibabacloud-tea-openapi
alibabacloud-tea-util
Pillow
markupsafe
imap-tools
